+++ settings.c
@@ -15,9 +15,9 @@ void	set_the_scene(t_scene *scene, double x)
 	matrix_type	normal[3];
 
 	// cicle
-	pos_c[0] = x;
+	pos_c[0] = -2;
 	pos_c[1] = 0;
-	pos_c[2] = 320;
+	pos_c[2] = 25;
 
 	color_c[0] = 255;
 	color_c[1] = 255;
@@ -25,16 +25,16 @@ void	set_the_scene(t_scene *scene, double x)
 
 	// plane
 	pos[0] = 0;
-	pos[1] = 0;
-	pos[2] = 225;
+	pos[1] = -3;
+	pos[2] = 0;
 
-	color[0] = 255;
-	color[1] = 0;
+	color[0] = 0;
+	color[1] = 255;
 	color[2] = 0;
 
-	normal[0] = 1;
-	normal[1] = 0;
-	normal[2] = -1;
+	normal[0] = 0;
+	normal[1] = 1;
+	normal[2] = 0;
 
 	// normal[0] = x;
 	// x -= 10;
@@ -42,14 +42,41 @@ void	set_the_scene(t_scene *scene, double x)
 	// x -= 10;
 	// normal[2] = x;
 
-	// scene->objs = init_object(PLANE, pos, color);
-	// scene->objs->plane = init_plane(normal);
+	scene->objs = init_object(PLANE, pos, color);
+	scene->objs->plane = init_plane(normal);
 
-	// scene->objs->next = init_object(CIRCLE, pos_c, color_c);
-	// scene->objs->next->circle = init_circle(55);
+	scene->objs->next = init_object(CIRCLE, pos_c, color_c);
+	scene->objs->next->circle = init_circle(3);
 
-	scene->objs = init_object(CIRCLE, pos_c, color);
-	scene->objs->circle = init_circle(25);
+	// plane
+	pos[0] = 0;
+	pos[1] = 0;
+	pos[2] = 50;
+
+	color[0] = 0;
+	color[1] = 0;
+	color[2] = 255;
+
+	normal[0] = 0;
+	normal[1] = 0;
+	normal[2] = -1;
+
+	scene->objs->next->next = init_object(PLANE, pos, color);
+	scene->objs->next->next->plane = init_plane(normal);
+
+	pos_c[0] = 1;
+	pos_c[1] = -1;
+	pos_c[2] = 15;
+
+	color_c[0] = 92;
+	color_c[1] = 179;
+	color_c[2] = 120;
+
+	scene->objs->next->next->next = init_object(CIRCLE, pos_c, color_c);
+	scene->objs->next->next->next->circle = init_circle(2);
+
+	// scene->objs = init_object(CIRCLE, pos_c, color);
+	// scene->objs->circle = init_circle(25);
 
 	// pos[0] = 100;
 	// pos[1] = 0;
@@ -61,7 +88,7 @@ void	set_the_scene(t_scene *scene, double x)
 	// scene->circles = init_circle( 100, 0, 150, 55 );
 	// scene->circles->next = init_circle(-50, 0, 150, 32);
 
-	scene->lights = init_light( 35, 35, 35, NULL);
+	scene->lights = init_light( 10, 28, 0, NULL);
 
 	// scene->lights = init_light( 35, 35, 35, NULL);
 	// scene->lights = init_light( -300, 100, 150, NULL);
