+++ vector.h
@@ -16,7 +16,7 @@ t_vector	*init_vector(t_matrix	*stuff);
 t_vector	*init_vector_intarr(matrix_type	array[3]);
 t_vector	*dup_vct(t_vector *vctr);
 
-matrix_type	*get_val(t_vector *vector);
+void		get_val(t_vector *vector, matrix_type store[3]);
 
 void		print_vector(t_vector	*stuff);
 void		free_vector(t_vector **stuff);
